requirements.txt
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.3
cramjam==2.11.0
et_xmlfile==2.0.0
fastparquet==2025.12.0
fsspec==2025.12.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.2
metatrader5==5.0.5430
narwhals==2.14.0
numpy==2.3.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
plotly==6.5.0
pyarrow==22.0.0
pycparser==2.23
pyparsing==3.3.1
PySocks==1.7.1
python-bcb==0.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
wsproto==1.3.2